requirements.txt
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.1
future==0.18.2
netmiko==3.4.0
ntc-templates==3.0.0
paramiko==2.8.1
pycparser==2.21
PyNaCl==1.4.0
pyserial==3.5
python-dotenv==0.19.2
scp==0.14.2
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
